+++ src/marine_qc/quality_control/__init__.py
@@ -13,7 +13,6 @@ from .buoy_tracking_qc import (
     do_sst_noisy_check,
     do_sst_start_tail_check,
 )
-from .multiple_checks import do_multiple_grouped_check, do_multiple_individual_check, do_multiple_sequential_check
 from .qc_grouped_reports import do_bayesian_buddy_check, do_mds_buddy_check
 from .qc_individual_reports import (
     do_climatology_check,
@@ -33,6 +32,7 @@ from .qc_individual_reports import (
     do_valid_value_clim_check,
     do_wind_consistency_check,
 )
+from .qc_multiple_checks import do_multiple_grouped_check, do_multiple_individual_check, do_multiple_sequential_check
 from .qc_sequential_reports import (
     do_few_check,
     do_iquam_track_check,
